Improve the Goal documentation and structure
