--- Aplib.Core/Heuristics.cs
@@ -1,7 +1,0 @@
-namespace Aplib.Core
-{
-	public class Heuristics
-	{
-		public float Distance;
-	}
-}
