--- Aplib.Core/Tactic.cs
@@ -1,7 +1,0 @@
-namespace Aplib.Core
-{
-	public abstract class Tactic
-	{
-		public abstract void IterateBdiCycle();
-	}
-}
